+++ tests/test_tasks.py
@@ -1,3 +1,4 @@
+import os
 import sys
 
 import mock
@@ -260,7 +261,7 @@ class DockerTasksTestCase(unittest.TestCase):
                 init_kwargs=dict(registry='host:5000'),
                 expected_calls=[
                     mock.call.local('docker pull test:latest', quiet=False, use_cache=True),
-                    mock.call.local('for image in $(docker images --filter "dangling=true" --quiet); do docker rmi "$image"; done'),
+                    mock.call.local('for img in $(docker images --filter "dangling=true" --quiet); do docker rmi "$img"; done'),
                     mock.call.local('docker tag test:latest host:5000/test:latest', use_cache=True),
                     mock.call.local('docker push host:5000/test:latest', quiet=False, use_cache=True),
                     mock.call.local('docker rmi host:5000/test:latest', use_cache=True),
@@ -277,7 +278,7 @@ class DockerTasksTestCase(unittest.TestCase):
                 init_kwargs=dict(registry='host:5000', ssh_tunnel_port=1234),
                 expected_calls=[
                     mock.call.local('docker pull test:latest', quiet=False, use_cache=True),
-                    mock.call.local('for image in $(docker images --filter "dangling=true" --quiet); do docker rmi "$image"; done'),
+                    mock.call.local('for img in $(docker images --filter "dangling=true" --quiet); do docker rmi "$img"; done'),
                     mock.call.local('docker tag test:latest host:5000/test:latest', use_cache=True),
                     mock.call.local('docker push host:5000/test:latest', quiet=False, use_cache=True),
                     mock.call.local('docker rmi host:5000/test:latest', use_cache=True),
@@ -320,7 +321,7 @@ class DockerTasksTestCase(unittest.TestCase):
                 init_kwargs=dict(registry='host:5000'),
                 expected_calls=[
                     mock.call.local('docker pull test:latest', quiet=False, use_cache=True),
-                    mock.call.local('for image in $(docker images --filter "dangling=true" --quiet); do docker rmi "$image"; done'),
+                    mock.call.local('for img in $(docker images --filter "dangling=true" --quiet); do docker rmi "$img"; done'),
                     mock.call.local('docker tag test:latest host:5000/test:latest', use_cache=True),
                     mock.call.local('docker push host:5000/test:latest', quiet=False, use_cache=True),
                     mock.call.local('docker rmi host:5000/test:latest', use_cache=True),
@@ -337,7 +338,7 @@ class DockerTasksTestCase(unittest.TestCase):
                 init_kwargs=dict(registry='host:5000', ssh_tunnel_port=1234),
                 expected_calls=[
                     mock.call.local('docker pull test:latest', quiet=False, use_cache=True),
-                    mock.call.local('for image in $(docker images --filter "dangling=true" --quiet); do docker rmi "$image"; done'),
+                    mock.call.local('for img in $(docker images --filter "dangling=true" --quiet); do docker rmi "$img"; done'),
                     mock.call.local('docker tag test:latest host:5000/test:latest', use_cache=True),
                     mock.call.local('docker push host:5000/test:latest', quiet=False, use_cache=True),
                     mock.call.local('docker rmi host:5000/test:latest', use_cache=True),
@@ -355,7 +356,7 @@ class DockerTasksTestCase(unittest.TestCase):
                 init_kwargs=dict(registry='host:4000'),
                 expected_calls=[
                     mock.call.local('docker pull registry:5000/test:latest', quiet=False, use_cache=True),
-                    mock.call.local('for image in $(docker images --filter "dangling=true" --quiet); do docker rmi "$image"; done'),
+                    mock.call.local('for img in $(docker images --filter "dangling=true" --quiet); do docker rmi "$img"; done'),
                     mock.call.local('docker tag registry:5000/test:latest host:4000/test:latest', use_cache=True),
                     mock.call.local('docker push host:4000/test:latest', quiet=False, use_cache=True),
                     mock.call.local('docker rmi host:4000/test:latest', use_cache=True),
@@ -372,7 +373,7 @@ class DockerTasksTestCase(unittest.TestCase):
                 init_kwargs=dict(registry='host:4000', ssh_tunnel_port=1234),
                 expected_calls=[
                     mock.call.local('docker pull registry:5000/test:latest', quiet=False, use_cache=True),
-                    mock.call.local('for image in $(docker images --filter "dangling=true" --quiet); do docker rmi "$image"; done'),
+                    mock.call.local('for img in $(docker images --filter "dangling=true" --quiet); do docker rmi "$img"; done'),
                     mock.call.local('docker tag registry:5000/test:latest host:4000/test:latest', use_cache=True),
                     mock.call.local('docker push host:4000/test:latest', quiet=False, use_cache=True),
                     mock.call.local('docker rmi host:4000/test:latest', use_cache=True),
@@ -460,7 +461,7 @@ class DockerTasksTestCase(unittest.TestCase):
                 init_kwargs=dict(registry='host:4000', ssh_tunnel_port=1234),
                 expected_calls=[
                     mock.call.local('docker pull registry:5000/test:tag', quiet=False, use_cache=True),
-                    mock.call.local('for image in $(docker images --filter "dangling=true" --quiet); do docker rmi "$image"; done'),
+                    mock.call.local('for img in $(docker images --filter "dangling=true" --quiet); do docker rmi "$img"; done'),
                     mock.call.local('docker tag registry:5000/test:tag host:4000/test:tag', use_cache=True),
                     mock.call.local('docker push host:4000/test:tag', quiet=False, use_cache=True),
                     mock.call.local('docker rmi host:4000/test:tag', use_cache=True),
@@ -563,6 +564,25 @@ class DockerTasksTestCase(unittest.TestCase):
                 fab.execute(commands.restore)
                 restore.assert_called_once()
 
+    def test_delete_dangling_images(self):
+        cases = dict(
+            windows=dict(
+                os_name='nt',
+                expected_command="for /F %i in ('docker images --filter \"dangling=true\" --quiet') do @docker rmi %i",
+            ),
+            posix=dict(
+                os_name='posix',
+                expected_command='for img in $(docker images --filter "dangling=true" --quiet); do docker rmi "$img"; done',
+            ),
+        )
+        for case, data in cases.items():
+            with self.subTest(case=case):
+                with mock.patch.object(fabricio, 'local') as local:
+                    with mock.patch('os.name', data['os_name']):
+                        tasks_list = tasks.DockerTasks(container=docker.Container(name='name'))
+                        tasks_list.delete_dangling_images()
+                        local.assert_called_once_with(data['expected_command'])
+
 
 class PullDockerTasksTestCase(unittest.TestCase):
 
@@ -584,7 +604,7 @@ class PullDockerTasksTestCase(unittest.TestCase):
                 deploy_kwargs=dict(),
                 expected_calls=[
                     mock.call.local('docker pull test:latest', quiet=False, use_cache=True),
-                    mock.call.local('for image in $(docker images --filter "dangling=true" --quiet); do docker rmi "$image"; done'),
+                    mock.call.local('for img in $(docker images --filter "dangling=true" --quiet); do docker rmi "$img"; done'),
                     mock.call.local('docker tag test:latest localhost:5000/test:latest', use_cache=True),
                     mock.call.local('docker push localhost:5000/test:latest', quiet=False, use_cache=True),
                     mock.call.local('docker rmi localhost:5000/test:latest', use_cache=True),
@@ -598,7 +618,7 @@ class PullDockerTasksTestCase(unittest.TestCase):
                 deploy_kwargs=dict(),
                 expected_calls=[
                     mock.call.local('docker pull test:latest', quiet=False, use_cache=True),
-                    mock.call.local('for image in $(docker images --filter "dangling=true" --quiet); do docker rmi "$image"; done'),
+                    mock.call.local('for img in $(docker images --filter "dangling=true" --quiet); do docker rmi "$img"; done'),
                     mock.call.local('docker tag test:latest localhost:5000/test:latest', use_cache=True),
                     mock.call.local('docker push localhost:5000/test:latest', quiet=False, use_cache=True),
                     mock.call.local('docker rmi localhost:5000/test:latest', use_cache=True),
@@ -612,7 +632,7 @@ class PullDockerTasksTestCase(unittest.TestCase):
                 deploy_kwargs=dict(),
                 expected_calls=[
                     mock.call.local('docker pull test:latest', quiet=False, use_cache=True),
-                    mock.call.local('for image in $(docker images --filter "dangling=true" --quiet); do docker rmi "$image"; done'),
+                    mock.call.local('for img in $(docker images --filter "dangling=true" --quiet); do docker rmi "$img"; done'),
                     mock.call.local('docker tag test:latest localhost:5000/test:latest', use_cache=True),
                     mock.call.local('docker push localhost:5000/test:latest', quiet=False, use_cache=True),
                     mock.call.local('docker rmi localhost:5000/test:latest', use_cache=True),
@@ -628,7 +648,7 @@ class PullDockerTasksTestCase(unittest.TestCase):
                 deploy_kwargs=dict(),
                 expected_calls=[
                     mock.call.local('docker pull test:latest', quiet=False, use_cache=True),
-                    mock.call.local('for image in $(docker images --filter "dangling=true" --quiet); do docker rmi "$image"; done'),
+                    mock.call.local('for img in $(docker images --filter "dangling=true" --quiet); do docker rmi "$img"; done'),
                     mock.call.local('docker tag test:latest localhost:5000/test:latest', use_cache=True),
                     mock.call.local('docker push localhost:5000/test:latest', quiet=False, use_cache=True),
                     mock.call.local('docker rmi localhost:5000/test:latest', use_cache=True),
@@ -645,7 +665,7 @@ class PullDockerTasksTestCase(unittest.TestCase):
                 deploy_kwargs=dict(),
                 expected_calls=[
                     mock.call.local('docker pull test:latest', quiet=False, use_cache=True),
-                    mock.call.local('for image in $(docker images --filter "dangling=true" --quiet); do docker rmi "$image"; done'),
+                    mock.call.local('for img in $(docker images --filter "dangling=true" --quiet); do docker rmi "$img"; done'),
                     mock.call.local('docker tag test:latest localhost:5000/test:latest', use_cache=True),
                     mock.call.local('docker push localhost:5000/test:latest', quiet=False, use_cache=True),
                     mock.call.local('docker rmi localhost:5000/test:latest', use_cache=True),
@@ -662,7 +682,7 @@ class PullDockerTasksTestCase(unittest.TestCase):
                 deploy_kwargs=dict(),
                 expected_calls=[
                     mock.call.local('docker pull test:latest', quiet=False, use_cache=True),
-                    mock.call.local('for image in $(docker images --filter "dangling=true" --quiet); do docker rmi "$image"; done'),
+                    mock.call.local('for img in $(docker images --filter "dangling=true" --quiet); do docker rmi "$img"; done'),
                     mock.call.local('docker tag test:latest host:1234/test:latest', use_cache=True),
                     mock.call.local('docker push host:1234/test:latest', quiet=False, use_cache=True),
                     mock.call.local('docker rmi host:1234/test:latest', use_cache=True),
@@ -677,7 +697,7 @@ class PullDockerTasksTestCase(unittest.TestCase):
                 deploy_kwargs=dict(force='yes'),
                 expected_calls=[
                     mock.call.local('docker pull test:latest', quiet=False, use_cache=True),
-                    mock.call.local('for image in $(docker images --filter "dangling=true" --quiet); do docker rmi "$image"; done'),
+                    mock.call.local('for img in $(docker images --filter "dangling=true" --quiet); do docker rmi "$img"; done'),
                     mock.call.local('docker tag test:latest localhost:5000/test:latest', use_cache=True),
                     mock.call.local('docker push localhost:5000/test:latest', quiet=False, use_cache=True),
                     mock.call.local('docker rmi localhost:5000/test:latest', use_cache=True),
@@ -691,7 +711,7 @@ class PullDockerTasksTestCase(unittest.TestCase):
                 deploy_kwargs=dict(tag='tag'),
                 expected_calls=[
                     mock.call.local('docker pull test:tag', quiet=False, use_cache=True),
-                    mock.call.local('for image in $(docker images --filter "dangling=true" --quiet); do docker rmi "$image"; done'),
+                    mock.call.local('for img in $(docker images --filter "dangling=true" --quiet); do docker rmi "$img"; done'),
                     mock.call.local('docker tag test:tag localhost:5000/test:tag', use_cache=True),
                     mock.call.local('docker push localhost:5000/test:tag', quiet=False, use_cache=True),
                     mock.call.local('docker rmi localhost:5000/test:tag', use_cache=True),
@@ -705,7 +725,7 @@ class PullDockerTasksTestCase(unittest.TestCase):
                 deploy_kwargs=dict(backup='yes'),
                 expected_calls=[
                     mock.call.local('docker pull test:latest', quiet=False, use_cache=True),
-                    mock.call.local('for image in $(docker images --filter "dangling=true" --quiet); do docker rmi "$image"; done'),
+                    mock.call.local('for img in $(docker images --filter "dangling=true" --quiet); do docker rmi "$img"; done'),
                     mock.call.local('docker tag test:latest localhost:5000/test:latest', use_cache=True),
                     mock.call.local('docker push localhost:5000/test:latest', quiet=False, use_cache=True),
                     mock.call.local('docker rmi localhost:5000/test:latest', use_cache=True),
@@ -720,7 +740,7 @@ class PullDockerTasksTestCase(unittest.TestCase):
                 deploy_kwargs=dict(migrate='no'),
                 expected_calls=[
                     mock.call.local('docker pull test:latest', quiet=False, use_cache=True),
-                    mock.call.local('for image in $(docker images --filter "dangling=true" --quiet); do docker rmi "$image"; done'),
+                    mock.call.local('for img in $(docker images --filter "dangling=true" --quiet); do docker rmi "$img"; done'),
                     mock.call.local('docker tag test:latest localhost:5000/test:latest', use_cache=True),
                     mock.call.local('docker push localhost:5000/test:latest', quiet=False, use_cache=True),
                     mock.call.local('docker rmi localhost:5000/test:latest', use_cache=True),
@@ -767,7 +787,7 @@ class BuildDockerTasksTestCase(unittest.TestCase):
                 deploy_kwargs=dict(),
                 expected_calls=[
                     mock.call.local('docker build --tag test:latest --pull .', quiet=False, use_cache=True),
-                    mock.call.local('for image in $(docker images --filter "dangling=true" --quiet); do docker rmi "$image"; done'),
+                    mock.call.local('for img in $(docker images --filter "dangling=true" --quiet); do docker rmi "$img"; done'),
                     mock.call.local('docker tag test:latest localhost:5000/test:latest', use_cache=True),
                     mock.call.local('docker push localhost:5000/test:latest', quiet=False, use_cache=True),
                     mock.call.local('docker rmi localhost:5000/test:latest', use_cache=True),
@@ -781,7 +801,7 @@ class BuildDockerTasksTestCase(unittest.TestCase):
                 deploy_kwargs=dict(),
                 expected_calls=[
                     mock.call.local('docker build --tag test:latest --pull .', quiet=False, use_cache=True),
-                    mock.call.local('for image in $(docker images --filter "dangling=true" --quiet); do docker rmi "$image"; done'),
+                    mock.call.local('for img in $(docker images --filter "dangling=true" --quiet); do docker rmi "$img"; done'),
                     mock.call.local('docker tag test:latest localhost:5000/test:latest', use_cache=True),
                     mock.call.local('docker push localhost:5000/test:latest', quiet=False, use_cache=True),
                     mock.call.local('docker rmi localhost:5000/test:latest', use_cache=True),
@@ -795,7 +815,7 @@ class BuildDockerTasksTestCase(unittest.TestCase):
                 deploy_kwargs=dict(no_cache=True),
                 expected_calls=[
                     mock.call.local('docker build --tag test:latest --no-cache --pull .', quiet=False, use_cache=True),
-                    mock.call.local('for image in $(docker images --filter "dangling=true" --quiet); do docker rmi "$image"; done'),
+                    mock.call.local('for img in $(docker images --filter "dangling=true" --quiet); do docker rmi "$img"; done'),
                     mock.call.local('docker tag test:latest localhost:5000/test:latest', use_cache=True),
                     mock.call.local('docker push localhost:5000/test:latest', quiet=False, use_cache=True),
                     mock.call.local('docker rmi localhost:5000/test:latest', use_cache=True),
@@ -809,7 +829,7 @@ class BuildDockerTasksTestCase(unittest.TestCase):
                 deploy_kwargs=dict(),
                 expected_calls=[
                     mock.call.local('docker build --tag test:latest --pull build/path', quiet=False, use_cache=True),
-                    mock.call.local('for image in $(docker images --filter "dangling=true" --quiet); do docker rmi "$image"; done'),
+                    mock.call.local('for img in $(docker images --filter "dangling=true" --quiet); do docker rmi "$img"; done'),
                     mock.call.local('docker tag test:latest localhost:5000/test:latest', use_cache=True),
                     mock.call.local('docker push localhost:5000/test:latest', quiet=False, use_cache=True),
                     mock.call.local('docker rmi localhost:5000/test:latest', use_cache=True),
@@ -824,7 +844,7 @@ class BuildDockerTasksTestCase(unittest.TestCase):
                 deploy_kwargs=dict(),
                 expected_calls=[
                     mock.call.local('docker build --tag test:latest --pull .', quiet=False, use_cache=True),
-                    mock.call.local('for image in $(docker images --filter "dangling=true" --quiet); do docker rmi "$image"; done'),
+                    mock.call.local('for img in $(docker images --filter "dangling=true" --quiet); do docker rmi "$img"; done'),
                     mock.call.local('docker tag test:latest localhost:5000/test:latest', use_cache=True),
                     mock.call.local('docker push localhost:5000/test:latest', quiet=False, use_cache=True),
                     mock.call.local('docker rmi localhost:5000/test:latest', use_cache=True),
@@ -840,7 +860,7 @@ class BuildDockerTasksTestCase(unittest.TestCase):
                 deploy_kwargs=dict(),
                 expected_calls=[
                     mock.call.local('docker build --tag test:latest --pull .', quiet=False, use_cache=True),
-                    mock.call.local('for image in $(docker images --filter "dangling=true" --quiet); do docker rmi "$image"; done'),
+                    mock.call.local('for img in $(docker images --filter "dangling=true" --quiet); do docker rmi "$img"; done'),
                     mock.call.local('docker tag test:latest localhost:5000/test:latest', use_cache=True),
                     mock.call.local('docker push localhost:5000/test:latest', quiet=False, use_cache=True),
                     mock.call.local('docker rmi localhost:5000/test:latest', use_cache=True),
@@ -857,7 +877,7 @@ class BuildDockerTasksTestCase(unittest.TestCase):
                 deploy_kwargs=dict(),
                 expected_calls=[
                     mock.call.local('docker build --tag test:latest --pull .', quiet=False, use_cache=True),
-                    mock.call.local('for image in $(docker images --filter "dangling=true" --quiet); do docker rmi "$image"; done'),
+                    mock.call.local('for img in $(docker images --filter "dangling=true" --quiet); do docker rmi "$img"; done'),
                     mock.call.local('docker tag test:latest localhost:5000/test:latest', use_cache=True),
                     mock.call.local('docker push localhost:5000/test:latest', quiet=False, use_cache=True),
                     mock.call.local('docker rmi localhost:5000/test:latest', use_cache=True),
@@ -874,7 +894,7 @@ class BuildDockerTasksTestCase(unittest.TestCase):
                 deploy_kwargs=dict(),
                 expected_calls=[
                     mock.call.local('docker build --tag test:latest --pull .', quiet=False, use_cache=True),
-                    mock.call.local('for image in $(docker images --filter "dangling=true" --quiet); do docker rmi "$image"; done'),
+                    mock.call.local('for img in $(docker images --filter "dangling=true" --quiet); do docker rmi "$img"; done'),
                     mock.call.local('docker tag test:latest host:1234/test:latest', use_cache=True),
                     mock.call.local('docker push host:1234/test:latest', quiet=False, use_cache=True),
                     mock.call.local('docker rmi host:1234/test:latest', use_cache=True),
@@ -889,7 +909,7 @@ class BuildDockerTasksTestCase(unittest.TestCase):
                 deploy_kwargs=dict(force='yes'),
                 expected_calls=[
                     mock.call.local('docker build --tag test:latest --pull .', quiet=False, use_cache=True),
-                    mock.call.local('for image in $(docker images --filter "dangling=true" --quiet); do docker rmi "$image"; done'),
+                    mock.call.local('for img in $(docker images --filter "dangling=true" --quiet); do docker rmi "$img"; done'),
                     mock.call.local('docker tag test:latest localhost:5000/test:latest', use_cache=True),
                     mock.call.local('docker push localhost:5000/test:latest', quiet=False, use_cache=True),
                     mock.call.local('docker rmi localhost:5000/test:latest', use_cache=True),
@@ -903,7 +923,7 @@ class BuildDockerTasksTestCase(unittest.TestCase):
                 deploy_kwargs=dict(tag='tag'),
                 expected_calls=[
                     mock.call.local('docker build --tag test:tag --pull .', quiet=False, use_cache=True),
-                    mock.call.local('for image in $(docker images --filter "dangling=true" --quiet); do docker rmi "$image"; done'),
+                    mock.call.local('for img in $(docker images --filter "dangling=true" --quiet); do docker rmi "$img"; done'),
                     mock.call.local('docker tag test:tag localhost:5000/test:tag', use_cache=True),
                     mock.call.local('docker push localhost:5000/test:tag', quiet=False, use_cache=True),
                     mock.call.local('docker rmi localhost:5000/test:tag', use_cache=True),
@@ -917,7 +937,7 @@ class BuildDockerTasksTestCase(unittest.TestCase):
                 deploy_kwargs=dict(backup='yes'),
                 expected_calls=[
                     mock.call.local('docker build --tag test:latest --pull .', quiet=False, use_cache=True),
-                    mock.call.local('for image in $(docker images --filter "dangling=true" --quiet); do docker rmi "$image"; done'),
+                    mock.call.local('for img in $(docker images --filter "dangling=true" --quiet); do docker rmi "$img"; done'),
                     mock.call.local('docker tag test:latest localhost:5000/test:latest', use_cache=True),
                     mock.call.local('docker push localhost:5000/test:latest', quiet=False, use_cache=True),
                     mock.call.local('docker rmi localhost:5000/test:latest', use_cache=True),
@@ -932,7 +952,7 @@ class BuildDockerTasksTestCase(unittest.TestCase):
                 deploy_kwargs=dict(migrate='no'),
                 expected_calls=[
                     mock.call.local('docker build --tag test:latest --pull .', quiet=False, use_cache=True),
-                    mock.call.local('for image in $(docker images --filter "dangling=true" --quiet); do docker rmi "$image"; done'),
+                    mock.call.local('for img in $(docker images --filter "dangling=true" --quiet); do docker rmi "$img"; done'),
                     mock.call.local('docker tag test:latest localhost:5000/test:latest', use_cache=True),
                     mock.call.local('docker push localhost:5000/test:latest', quiet=False, use_cache=True),
                     mock.call.local('docker rmi localhost:5000/test:latest', use_cache=True),
@@ -980,7 +1000,7 @@ class ImageBuildDockerTasksTestCase(unittest.TestCase):
                 init_kwargs=dict(),
                 expected_calls=[
                     mock.call.local('docker build --tag test:latest --pull .', quiet=False, use_cache=True),
-                    mock.call.local('for image in $(docker images --filter "dangling=true" --quiet); do docker rmi "$image"; done'),
+                    mock.call.local('for img in $(docker images --filter "dangling=true" --quiet); do docker rmi "$img"; done'),
                     mock.call.local('docker push test:latest', quiet=False, use_cache=True),
                     mock.call.run('docker pull test:latest', quiet=False),
                     mock.call.migrate(tag=None),
@@ -993,7 +1013,7 @@ class ImageBuildDockerTasksTestCase(unittest.TestCase):
                 init_kwargs=dict(),
                 expected_calls=[
                     mock.call.local('docker build --tag registry:5000/test:latest --pull .', quiet=False, use_cache=True),
-                    mock.call.local('for image in $(docker images --filter "dangling=true" --quiet); do docker rmi "$image"; done'),
+                    mock.call.local('for img in $(docker images --filter "dangling=true" --quiet); do docker rmi "$img"; done'),
                     mock.call.local('docker push registry:5000/test:latest', quiet=False, use_cache=True),
                     mock.call.run('docker pull registry:5000/test:latest', quiet=False),
                     mock.call.migrate(tag=None),
@@ -1006,7 +1026,7 @@ class ImageBuildDockerTasksTestCase(unittest.TestCase):
                 init_kwargs=dict(ssh_tunnel_port=1234),
                 expected_calls=[
                     mock.call.local('docker build --tag registry:5000/test:latest --pull .', quiet=False, use_cache=True),
-                    mock.call.local('for image in $(docker images --filter "dangling=true" --quiet); do docker rmi "$image"; done'),
+                    mock.call.local('for img in $(docker images --filter "dangling=true" --quiet); do docker rmi "$img"; done'),
                     mock.call.local('docker push registry:5000/test:latest', quiet=False, use_cache=True),
                     mock.call.remote_tunnel(remote_port=1234, local_port=5000, local_host='registry'),
                     mock.call.run('docker pull localhost:1234/test:latest', quiet=False),
@@ -1022,7 +1042,7 @@ class ImageBuildDockerTasksTestCase(unittest.TestCase):
                 init_kwargs=dict(registry='host:5000'),
                 expected_calls=[
                     mock.call.local('docker build --tag host:5000/test:latest --pull .', quiet=False, use_cache=True),
-                    mock.call.local('for image in $(docker images --filter "dangling=true" --quiet); do docker rmi "$image"; done'),
+                    mock.call.local('for img in $(docker images --filter "dangling=true" --quiet); do docker rmi "$img"; done'),
                     mock.call.local('docker push host:5000/test:latest', quiet=False, use_cache=True),
                     mock.call.run('docker pull host:5000/test:latest', quiet=False),
                     mock.call.run('docker tag host:5000/test:latest test:latest'),
@@ -1037,7 +1057,7 @@ class ImageBuildDockerTasksTestCase(unittest.TestCase):
                 init_kwargs=dict(registry='host:5000', ssh_tunnel_port=1234),
                 expected_calls=[
                     mock.call.local('docker build --tag host:5000/test:latest --pull .', quiet=False, use_cache=True),
-                    mock.call.local('for image in $(docker images --filter "dangling=true" --quiet); do docker rmi "$image"; done'),
+                    mock.call.local('for img in $(docker images --filter "dangling=true" --quiet); do docker rmi "$img"; done'),
                     mock.call.local('docker push host:5000/test:latest', quiet=False, use_cache=True),
                     mock.call.remote_tunnel(remote_port=1234, local_port=5000, local_host='host'),
                     mock.call.run('docker pull localhost:1234/test:latest', quiet=False),
@@ -1053,7 +1073,7 @@ class ImageBuildDockerTasksTestCase(unittest.TestCase):
                 init_kwargs=dict(registry='host:4000'),
                 expected_calls=[
                     mock.call.local('docker build --tag host:4000/test:latest --pull .', quiet=False, use_cache=True),
-                    mock.call.local('for image in $(docker images --filter "dangling=true" --quiet); do docker rmi "$image"; done'),
+                    mock.call.local('for img in $(docker images --filter "dangling=true" --quiet); do docker rmi "$img"; done'),
                     mock.call.local('docker push host:4000/test:latest', quiet=False, use_cache=True),
                     mock.call.run('docker pull host:4000/test:latest', quiet=False),
                     mock.call.run('docker tag host:4000/test:latest registry:5000/test:latest'),
@@ -1068,7 +1088,7 @@ class ImageBuildDockerTasksTestCase(unittest.TestCase):
                 init_kwargs=dict(registry='host:4000', ssh_tunnel_port=1234),
                 expected_calls=[
                     mock.call.local('docker build --tag host:4000/test:latest --pull .', quiet=False, use_cache=True),
-                    mock.call.local('for image in $(docker images --filter "dangling=true" --quiet); do docker rmi "$image"; done'),
+                    mock.call.local('for img in $(docker images --filter "dangling=true" --quiet); do docker rmi "$img"; done'),
                     mock.call.local('docker push host:4000/test:latest', quiet=False, use_cache=True),
                     mock.call.remote_tunnel(remote_port=1234, local_port=4000, local_host='host'),
                     mock.call.run('docker pull localhost:1234/test:latest', quiet=False),
@@ -1084,7 +1104,7 @@ class ImageBuildDockerTasksTestCase(unittest.TestCase):
                 init_kwargs=dict(),
                 expected_calls=[
                     mock.call.local('docker build --tag test:latest --pull .', quiet=False, use_cache=True),
-                    mock.call.local('for image in $(docker images --filter "dangling=true" --quiet); do docker rmi "$image"; done'),
+                    mock.call.local('for img in $(docker images --filter "dangling=true" --quiet); do docker rmi "$img"; done'),
                     mock.call.local('docker push test:latest', quiet=False, use_cache=True),
                     mock.call.run('docker pull test:latest', quiet=False),
                     mock.call.migrate(tag=None),
@@ -1097,7 +1117,7 @@ class ImageBuildDockerTasksTestCase(unittest.TestCase):
                 init_kwargs=dict(),
                 expected_calls=[
                     mock.call.local('docker build --tag test:latest --pull .', quiet=False, use_cache=True),
-                    mock.call.local('for image in $(docker images --filter "dangling=true" --quiet); do docker rmi "$image"; done'),
+                    mock.call.local('for img in $(docker images --filter "dangling=true" --quiet); do docker rmi "$img"; done'),
                     mock.call.local('docker push test:latest', quiet=False, use_cache=True),
                     mock.call.run('docker pull test:latest', quiet=False),
                     mock.call.migrate(tag=None),
@@ -1110,7 +1130,7 @@ class ImageBuildDockerTasksTestCase(unittest.TestCase):
                 init_kwargs=dict(build_path='foo'),
                 expected_calls=[
                     mock.call.local('docker build --tag test:latest --pull foo', quiet=False, use_cache=True),
-                    mock.call.local('for image in $(docker images --filter "dangling=true" --quiet); do docker rmi "$image"; done'),
+                    mock.call.local('for img in $(docker images --filter "dangling=true" --quiet); do docker rmi "$img"; done'),
                     mock.call.local('docker push test:latest', quiet=False, use_cache=True),
                     mock.call.run('docker pull test:latest', quiet=False),
                     mock.call.migrate(tag=None),
@@ -1123,7 +1143,7 @@ class ImageBuildDockerTasksTestCase(unittest.TestCase):
                 init_kwargs=dict(),
                 expected_calls=[
                     mock.call.local('docker build --tag test:tag --pull .', quiet=False, use_cache=True),
-                    mock.call.local('for image in $(docker images --filter "dangling=true" --quiet); do docker rmi "$image"; done'),
+                    mock.call.local('for img in $(docker images --filter "dangling=true" --quiet); do docker rmi "$img"; done'),
                     mock.call.local('docker push test:tag', quiet=False, use_cache=True),
                     mock.call.run('docker pull test:tag', quiet=False),
                     mock.call.migrate(tag='tag'),
@@ -1136,7 +1156,7 @@ class ImageBuildDockerTasksTestCase(unittest.TestCase):
                 init_kwargs=dict(),
                 expected_calls=[
                     mock.call.local('docker build --tag test:latest --pull .', quiet=False, use_cache=True),
-                    mock.call.local('for image in $(docker images --filter "dangling=true" --quiet); do docker rmi "$image"; done'),
+                    mock.call.local('for img in $(docker images --filter "dangling=true" --quiet); do docker rmi "$img"; done'),
                     mock.call.local('docker push test:latest', quiet=False, use_cache=True),
                     mock.call.backup(),
                     mock.call.run('docker pull test:latest', quiet=False),
@@ -1150,7 +1170,7 @@ class ImageBuildDockerTasksTestCase(unittest.TestCase):
                 init_kwargs=dict(),
                 expected_calls=[
                     mock.call.local('docker build --tag test:latest --pull .', quiet=False, use_cache=True),
-                    mock.call.local('for image in $(docker images --filter "dangling=true" --quiet); do docker rmi "$image"; done'),
+                    mock.call.local('for img in $(docker images --filter "dangling=true" --quiet); do docker rmi "$img"; done'),
                     mock.call.local('docker push test:latest', quiet=False, use_cache=True),
                     mock.call.run('docker pull test:latest', quiet=False),
                     mock.call.migrate(tag=None),
@@ -1163,7 +1183,7 @@ class ImageBuildDockerTasksTestCase(unittest.TestCase):
                 init_kwargs=dict(),
                 expected_calls=[
                     mock.call.local('docker build --tag test:latest --pull .', quiet=False, use_cache=True),
-                    mock.call.local('for image in $(docker images --filter "dangling=true" --quiet); do docker rmi "$image"; done'),
+                    mock.call.local('for img in $(docker images --filter "dangling=true" --quiet); do docker rmi "$img"; done'),
                     mock.call.local('docker push test:latest', quiet=False, use_cache=True),
                     mock.call.run('docker pull test:latest', quiet=False),
                     mock.call.update(force=False, tag=None),
@@ -1175,7 +1195,7 @@ class ImageBuildDockerTasksTestCase(unittest.TestCase):
                 init_kwargs=dict(),
                 expected_calls=[
                     mock.call.local('docker build --tag test:latest --pull .', quiet=False, use_cache=True),
-                    mock.call.local('for image in $(docker images --filter "dangling=true" --quiet); do docker rmi "$image"; done'),
+                    mock.call.local('for img in $(docker images --filter "dangling=true" --quiet); do docker rmi "$img"; done'),
                     mock.call.local('docker push test:latest', quiet=False, use_cache=True),
                     mock.call.run('docker pull test:latest', quiet=False),
                     mock.call.migrate(tag=None),
@@ -1198,7 +1218,7 @@ class ImageBuildDockerTasksTestCase(unittest.TestCase):
                 init_kwargs=dict(),
                 expected_calls=[
                     mock.call.local('docker build --tag test:latest --pull .', quiet=False, use_cache=True),
-                    mock.call.local('for image in $(docker images --filter "dangling=true" --quiet); do docker rmi "$image"; done'),
+                    mock.call.local('for img in $(docker images --filter "dangling=true" --quiet); do docker rmi "$img"; done'),
                     mock.call.local('docker push test:latest', quiet=False, use_cache=True),
                     mock.call.run('docker pull test:latest', quiet=False),
                     mock.call.migrate(tag=None),
@@ -1211,7 +1231,7 @@ class ImageBuildDockerTasksTestCase(unittest.TestCase):
                 init_kwargs=dict(registry='host:4000', ssh_tunnel_port=1234, build_path='foo'),
                 expected_calls=[
                     mock.call.local('docker build --tag host:4000/test:tag --pull foo', quiet=False, use_cache=True),
-                    mock.call.local('for image in $(docker images --filter "dangling=true" --quiet); do docker rmi "$image"; done'),
+                    mock.call.local('for img in $(docker images --filter "dangling=true" --quiet); do docker rmi "$img"; done'),
                     mock.call.local('docker push host:4000/test:tag', quiet=False, use_cache=True),
                     mock.call.backup(),
                     mock.call.remote_tunnel(remote_port=1234, local_port=4000, local_host='host'),
@@ -1252,21 +1272,21 @@ class ImageBuildDockerTasksTestCase(unittest.TestCase):
                 kwargs=dict(),
                 expected_calls=[
                     mock.call('docker build --tag image:latest --pull .', quiet=False, use_cache=True),
-                    mock.call.local('for image in $(docker images --filter "dangling=true" --quiet); do docker rmi "$image"; done'),
+                    mock.call.local('for img in $(docker images --filter "dangling=true" --quiet); do docker rmi "$img"; done'),
                 ]
             ),
             explicit_cache=dict(
                 kwargs=dict(no_cache='no'),
                 expected_calls=[
                     mock.call('docker build --tag image:latest --pull .', quiet=False, use_cache=True),
-                    mock.call.local('for image in $(docker images --filter "dangling=true" --quiet); do docker rmi "$image"; done'),
+                    mock.call.local('for img in $(docker images --filter "dangling=true" --quiet); do docker rmi "$img"; done'),
                 ]
             ),
             no_cache=dict(
                 kwargs=dict(no_cache='yes'),
                 expected_calls=[
                     mock.call('docker build --tag image:latest --no-cache --pull .', quiet=False, use_cache=True),
-                    mock.call.local('for image in $(docker images --filter "dangling=true" --quiet); do docker rmi "$image"; done'),
+                    mock.call.local('for img in $(docker images --filter "dangling=true" --quiet); do docker rmi "$img"; done'),
                 ]
             ),
         )